Created verifier module
Comparison with RNAStrands and mmcif files to verify the data.
The Venn diagram in db_overlap shws overlap of RNALoops and RNAStrands.
The RNALoops_vs_RNAStrands file shows the comparsion.

The verify package might be used with the main_verify.py function.
It creates the quality.csv files to get multiloop qualities.
Based on that the barplot with the entries per quality is generated.