Added engineer package
engineer package contains all modules and functions to create and work with
aggregated data (i.e. multiloops aggregated by secondary structure).
It has the add_features module to aggregate the raw df and add features like
mean, median, std of planar angles of the groups. Also some more features
about the chains etc. can be added, check the modules for details.

analyse_cluster is a massive module to analyse clusters obtained by the
clustering package. It can create visualizations of many relevant characteristics
of the clusters obtained and cut them in a desired manner. Check the module for
details.

predict_prep is not used at the moment, but might be relevant later.