Added evaluater package
evaluater can be used to analyse clusters using ellbow plots, silhouette analysis
and other techniques. It is not required by any other module or main module at
the moment, but can be exploited, if the quality of clusters should be analysed
or the optimum number of clusters should be found.